查看用户图集

+++ src/main/java/com/xinshijie/gallery/mapper/UserAlbumMapper.java
@@ -48,12 +48,12 @@ public interface UserAlbumMapper extends BaseMapper<UserAlbum> {
     /**
      * 删除数据
      */
-    Integer delById(Integer userId, Long id);
+    Integer delById(Integer userId, Integer id);
 
     /**
      * 根据id数据
      */
-    UserAlbumVo getInfo(Integer userId, Long id);
+    UserAlbumVo getInfo(Integer userId, Integer id);
 
     Integer updateCountImage(Integer id);
